% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.add_parenth}
\alias{.add_parenth}
\title{Add parentheses around a string}
\usage{
.add_parenth(str, indent = TRUE)
}
\arguments{
\item{str}{a single string, often \code{sql} objects.}

\item{indent}{a logical value, decides whether indentation should match the
newly created parentheses.}
}
\value{
The \code{str} string wrapped with parentheses.
}
\description{
Meant for modifying \code{sql} objects safely (preserving attributes) and wrapping
them in parentheses additionally indenting where necessary.
}
\keyword{internal}
