% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.align_lines}
\alias{.align_lines}
\title{Align lengths of the last lines to the longest string}
\usage{
.align_lines(strings)
}
\arguments{
\item{strings}{a character vector of strings to align.}
}
\value{
The character vectors with spaces added at the end (if necessary).
}
\description{
This function adds spaces at the end of the last line of every string
in \code{strings}, so that the last line has the same amount of characters
as the longest line in ALL strings in \code{strings}.
This is useful for aligning "AS" statements, e.g. in SELECT.
}
\details{
Significance of aligning the last stems from possible multiline expressions,
possibly not strictly necessary in most cases.
}
\keyword{internal}
