% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns_picker.R
\name{.columns_formatter}
\alias{.columns_formatter}
\title{Internal columns formatter}
\usage{
.columns_formatter(cols)
}
\arguments{
\item{cols}{a character vector of columns. Vector names may be used to provide
aliases to columns.}
}
\value{
A formatted string of columns.
}
\description{
Formats a character vector of provided columns (e.g. in SELECT, ORDER BY,
GROUP BY...). This aligns "AS" statements in one line for pretty visual
output.
}
\keyword{internal}
