% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defusing.R
\name{.defuse_calls}
\alias{.defuse_calls}
\title{Defuse internal calls}
\usage{
.defuse_calls(quo)
}
\arguments{
\item{quo}{quosure to defuse. Should be a call, see \code{\link[rlang:is_call]{rlang::is_call()}}.}
}
\value{
The quosure with a modified expression.
}
\description{
Defuse internal calls
}
\keyword{internal}
