% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.main_class}
\alias{.main_class}
\title{Main class of an object}
\usage{
.main_class(x)
}
\arguments{
\item{x}{an object to extract the class from.}
}
\value{
The first class of the object, as given by \code{class()}.
}
\description{
Main class of an object
}
\keyword{internal}
