% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_in.R
\name{.parse.sql_in}
\alias{.parse.sql_in}
\title{Internal parser for \code{sql_in} objects}
\usage{
\method{.parse}{sql_in}(x, fields, ...)
}
\arguments{
\item{x}{a \code{sql} object or a single string.}

\item{fields}{a list passed to \code{.parse} class methods. Internally set
within \code{.sql_parse()} wrapper to \code{attributes(x)}. The attributes
are set within object constructors.}

\item{...}{another arguments passed to \code{.parse()} and corresponding methods.}
}
\description{
Internal parser for \code{sql_in} objects
}
\keyword{internal}
