% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_tuple.R
\name{.parse.sql_tuple}
\alias{.parse.sql_tuple}
\title{Internal parser for \code{sql_tuple} objects}
\usage{
\method{.parse}{sql_tuple}(x, fields, as_values = FALSE, ...)
}
\arguments{
\item{x}{a \code{sql} object or a single string.}

\item{fields}{a list passed to \code{.parse} class methods. Internally set
within \code{.sql_parse()} wrapper to \code{attributes(x)}. The attributes
are set within object constructors.}

\item{as_values}{a logical value. If \code{TRUE}, then every element of the tuple
is considered as inner list of values, i.e. the entire vector is not wrapped
in outside parentheses and every item is on a new line. Used in \code{\link[=sql_insert]{sql_insert()}}.}

\item{...}{another arguments passed to \code{.parse()} and corresponding methods.}
}
\description{
Internal parser for \code{sql_tuple} objects
}
\keyword{internal}
