% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_logical.R
\name{.parse_logical}
\alias{.parse_logical}
\title{Title}
\usage{
.parse_logical(
  x,
  keyword,
  level = 0,
  add_parenth = FALSE,
  add_parenth_x = FALSE,
  add_parenth_y = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\keyword{internal}
