% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.prepare_vector}
\alias{.prepare_vector}
\title{Prepare a vector}
\usage{
.prepare_vector(x, short = FALSE)
}
\arguments{
\item{x}{an atomic vector to align.}

\item{short}{a logical value, whether this is meant to be as short as possible.
That means trimming whitespaces and dropping trailing zeros. Short version
is used in tuples, whereas longer in INSERTs.}
}
\value{
A character vector based on \code{x}.
}
\description{
This prepares atomic vectors to be used e.g. in tuples or INSERT statements.
Numerics are always converted to non-scientific version, with precision as long
as possible. \code{NA} values are replaced with "NULL" strings.
If necessary, alignment is applied, i.e. numerics to the right, characters to
the left, with constant length, so that the output is human readable.
}
\keyword{internal}
