% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.rename_call}
\alias{.rename_call}
\title{Change the name of a call without modifying arguments}
\usage{
.rename_call(call, name)
}
\arguments{
\item{call}{a call to be modified.}

\item{name}{a single string, new name for the call.}
}
\value{
The call with a modified name.
}
\description{
Change the name of a call without modifying arguments
}
\keyword{internal}
