% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_logical.R
\name{.sql_parse.sql_logical}
\alias{.sql_parse.sql_logical}
\title{Title}
\usage{
\method{.sql_parse}{sql_logical}(x, break_lines = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\keyword{internal}
