% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_prepare.R
\name{.sql_prepare}
\alias{.sql_prepare}
\title{Internal SQL preparation helper}
\usage{
.sql_prepare(..., defuse = TRUE, sql_like = TRUE, keep = "sql")
}
\arguments{
\item{...}{character vectors, names of columns. May also include other \code{sql} objects (e.g.\code{\link[=sql_vars]{sql_vars()}}, \code{\link[=sql_case]{sql_case()}}). Non-character elements are dropped with a warning, after dropping the vector cannot be empty. Vectors may be named to provide column aliases.}

\item{defuse}{a logical value, decides whether inputs in \code{...} should be
defused. This includes replacing shortened calls to \verb{sql_*()} functions
with full function names.}

\item{sql_like}{a logical value, if \code{TRUE} it is required that elements
in \code{...} are single strings or \code{sql} objects of classes listed in \code{keep}.
Otherwise, all character vectors are allowed in \code{...}.}

\item{keep}{a character vector of class names, indicates allowed classes of
objects to keep in the output if \code{sql_like = TRUE}.}
}
\value{
A list of prepared expressions/objects.
}
\description{
Internal SQL preparation helper
}
\keyword{internal}
