% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_tools.R
\name{is_sql}
\alias{is_sql}
\title{Check whether object is a \code{sql} object}
\usage{
is_sql(x)
}
\arguments{
\item{x}{Any object to be checked for \code{sql} class.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
\code{is_sql()} checks whether an object inherits from the general \code{sql} class.
}
