% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_methods.R
\name{print.sql}
\alias{print.sql}
\title{Print SQL objects}
\usage{
\method{print}{sql}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns \code{x} invisibly.
}
\description{
DESCRIPTION TO BE WRITTEN.
}
