% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_case.R
\name{sql_case}
\alias{sql_case}
\title{Construct a SQL CASE WHEN statement}
\usage{
sql_case(..., .else = NULL)
}
\arguments{
\item{...}{an atomic vector of conditions and values. Must be of even length.
Represents a sequence of \code{when1}, \code{then1}, \code{when2}, \code{then2} and so on...}

\item{.else}{optional scalar value. If not \code{NULL}, then the provided value
is used for the finale ELSE statement.}
}
\value{
A character string representing the SQL CASE WHEN statement, with S3
class 'sql_case'.
}
\description{
DESCRIPTION TO BE WRITTEN
}
