% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_condition.R
\name{sql_condition}
\alias{sql_condition}
\title{Create an SQL condition}
\usage{
sql_condition(..., .defuse = TRUE)
}
\arguments{
\item{...}{Arguments to be processed, representing conditions.}

\item{.defuse}{a logical value. Decides whether input arguments in \code{...}
should be defused. For more information about defusing, see \link{sql_defusing}.}
}
\value{
A character string representing the logical conditions, with S3
class 'sql_logical'.
}
\description{
TO BE WRITTEN
}
