% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_create_table.R
\name{sql_create_table}
\alias{sql_create_table}
\title{Construct a SQL CREATE TABLE statement}
\usage{
sql_create_table(name, variables = NULL, query = NULL)
}
\arguments{
\item{name}{a single string with the name of the table to be created.}

\item{variables}{a named character vector. Vector names represent names of
the columns, whereas vector values represent column types (e.g. integer).
Names must be unique.}

\item{query}{an \code{\link[=sql_query]{sql_query()}} object, query used as a basis
for the new table.}
}
\value{
A character string representing the SQL CREATE TABLE statement, with S3
class 'sql_create_table'.
}
\description{
Builds a CREATE TABLE statement. New tables can be defined by providing
column names and their types, or created from the results of a SELECT
query using \verb{CREATE TABLE ... AS}.
}
\examples{
sql_create_table("new_table",
                 variables = c(col1 = "integer", col2 = "text"))
q <- sql(select("*"),
         from("original"))
sql_create_table("new_table", query = q)

}
\seealso{
SQL data manipulation functions:
\code{\link[=sql_delete]{sql_delete()}},
\code{\link[=sql_insert]{sql_insert()}},
\code{\link[=sql_update]{sql_update()}}
}
\concept{dml}
