% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defusing.R
\name{sql_defusing}
\alias{sql_defusing}
\title{SQL defusing}
\description{
In \code{sqlqueries} package, defusing is used to simplify writing queries using
the power of Non-Standard Evaluation (NSE). This allows for:
\itemize{
\item using only keywords instead defused environments,
e.g. \code{sql(select("*"))} instead of \code{sql(sql_select("*"))},
\item wrapping elements of queries in parentheses preserves them in the query,
e.g. \code{sql_condition(("a" \%OR\% "b") \%AND\% "c")},
\item converting unevaluated expressions to strings like in \code{\link[=sql_vars]{sql_vars()}}.
}
}
\details{
In case one wants to use e.g. \code{select()} from another package or does not want
to rely on NSE, \code{.defuse} argument may be set to \code{FALSE} to disable this
behavior.
Defusing is turned on by default in all functions, as this significantly shortens
writing longer queries with several keywords.
}
