% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_group_by.R
\name{sql_group_by}
\alias{sql_group_by}
\title{Construct a SQL GROUP BY statement}
\usage{
sql_group_by(..., .defuse = TRUE)
}
\arguments{
\item{...}{character vectors, columns to group by. May also include other \code{sql} objects (e.g.\code{\link[=sql_vars]{sql_vars()}}, \code{\link[=sql_case]{sql_case()}}). Non-character elements are dropped with a warning, after dropping the vector cannot be empty.}

\item{.defuse}{a logical value. Decides whether input arguments in \code{...}
should be defused. For more information about defusing, see \link{sql_defusing}.}
}
\value{
A character string representing the SQL GROUP BY statement, with S3
class 'sql_group_by'.
}
\description{
Groups rows by the specified columns for use with aggregate functions.
Multiple columns can be provided and are included in the GROUP BY clause
in the order given.
}
\examples{
sql_group_by("a")
sql_group_by("a", "b")

}
\seealso{
SQL query statements:
\code{\link[=sql_from]{sql_from()}},
\code{\link[=sql_having]{sql_having()}},
\code{\link[=sql_join]{sql_join()}},
\code{\link[=sql_order_by]{sql_order_by()}},
\code{\link[=sql_select]{sql_select()}},
\code{\link[=sql_where]{sql_where()}}
}
\concept{building_blocks}
