% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_having.R
\name{sql_having}
\alias{sql_having}
\title{Construct a SQL HAVING statement}
\usage{
sql_having(..., .defuse = TRUE)
}
\arguments{
\item{...}{logical conditions to include. Can be character strings
or results of other \code{sql} objects (e.g. \code{\link[=sql_or]{sql_or()}}).}

\item{.defuse}{a logical value. Decides whether input arguments in \code{...}
should be defused. For more information about defusing, see \link{sql_defusing}.}
}
\value{
A character string representing the SQL HAVING statement, with S3
class 'sql_having'.
}
\description{
Filters grouped rows after a GROUP BY clause. Accepts aggregate
conditions that are combined with AND, supporting the same logical
composition patterns as \code{\link[=sql_where]{sql_where()}}.
}
\examples{
sql_having("s > 3")
sql_having("s > 3", "t > 5")

}
\seealso{
SQL query statements:
\code{\link[=sql_from]{sql_from()}},
\code{\link[=sql_group_by]{sql_group_by()}},
\code{\link[=sql_join]{sql_join()}},
\code{\link[=sql_order_by]{sql_order_by()}},
\code{\link[=sql_select]{sql_select()}},
\code{\link[=sql_where]{sql_where()}}
}
\concept{building_blocks}
