% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_join.R
\name{sql_join}
\alias{sql_join}
\title{Construct a SQL JOIN statement}
\usage{
sql_join(table, on, type = "inner", alias = NULL)
}
\arguments{
\item{table}{a single, non-empty string with the name of the table to join.}

\item{on}{a single, non-empty string with a condition to join on.}

\item{type}{a single string, the type of join. Possible values are:
\code{"inner"}, \code{"left"}, \code{"right"}, \code{"full"}, \code{"key"}).}

\item{alias}{a single string or \code{NULL}. If provided, serves as an alias for
the joined table.}
}
\value{
A character string representing the SQL JOIN statement, with S3
class 'sql_join'.
}
\description{
DESCRIPTION TO BE WRITTEN
}
