% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_logical.R
\name{sql_logical}
\alias{sql_logical}
\alias{sql_condition}
\alias{sql_and}
\alias{sql_or}
\alias{\%AND\%}
\alias{\%OR\%}
\title{Create logical SQL conditions}
\usage{
sql_logical(x, y, operator)

sql_condition(..., .defuse = TRUE)

sql_and(x, y)

sql_or(x, y)

x \%AND\% y

x \%OR\% y
}
\arguments{
\item{x, y}{non-empty strings, input arguments for the logical operator.}

\item{operator}{a non-empty string, name of the logical operator to be
used (e.g. \code{"and"} or \code{"or"}).}

\item{...}{Arguments to be processed, representing conditions.}

\item{.defuse}{a logical value. Decides whether input arguments in \code{...}
should be defused. For more information about defusing, see \link{sql_defusing}.}
}
\value{
A character string representing the logical conditions, with S3
class 'sql_logical'.
}
\description{
\code{sql_logical()} serves as a template for creating logical conditions in a
composable manner, e.g. AND or OR.
}
\examples{
sql_logical("x = 1", "y = 1", "or")
sql_and("x = 1", "y = 2")
sql_or("z = 2", "w = 3")
"x = 1" \%AND\% "y = 2"
"z = 2" \%OR\% "w = 3"
sql_condition("x = 1", "y = 2")

}
