% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_order_by.R
\name{sql_order_by}
\alias{sql_order_by}
\title{Construct a SQL ORDER BY statement}
\usage{
sql_order_by(..., .defuse = TRUE)
}
\arguments{
\item{...}{character vectors, columns to order by. May also include other \code{sql} objects (e.g.\code{\link[=sql_vars]{sql_vars()}}, \code{\link[=sql_case]{sql_case()}}). Non-character elements are dropped with a warning, after dropping the vector cannot be empty.}

\item{.defuse}{a logical value. Decides whether input arguments in \code{...}
should be defused. For more information about defusing, see \link{sql_defusing}.}
}
\value{
A character string representing the SQL ORDER BY statement, with S3
class 'sql_order_by'.
}
\description{
DESCRIPTION TO BE WRITTEN
}
