% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_parse.R
\name{sql_parse}
\alias{sql_parse}
\alias{.sql_parse}
\alias{.parse}
\title{Parsing \code{sql} objects into text}
\usage{
.sql_parse(x, ...)

.parse(x, fields, ...)
}
\arguments{
\item{x}{a \code{sql} object.}

\item{...}{another arguments passed to \code{.parse()} and corresponding methods.}

\item{fields}{a list passed to \code{.parse} class methods. Internally set
within \code{.sql_parse()} wrapper to \code{attributes(x)}. The attributes
are set within object constructors.}
}
\value{
A string with a parsed object --- same attributes, but with formatted
text.
}
\description{
The function uses \verb{sql_*} classes methods for \code{.parse} generic to transform
all the necessary attributes of an object (i.e. data parsed when constructing)
to create a formatted (parsed) text of the statement/query.
}
\keyword{internal}
