% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_update.R
\name{sql_update}
\alias{sql_update}
\title{Construct a SQL UPDATE statement}
\usage{
sql_update(table, set)
}
\arguments{
\item{table}{a single string with the name of the table to update.}

\item{set}{a list, every element must be named (with unique names) and
must contain only a scalar value.}
}
\value{
A character string representing the SQL UPDATE statement, with S3
class 'sql_update'.
}
\description{
DESCRIPTION TO BE WRITTEN
}
