% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlqueries-package.R
\docType{package}
\name{sqlqueries-package}
\alias{sqlqueries}
\alias{sqlqueries-package}
\title{sqlqueries: Programmatic SQL Query Construction}
\description{
A functional framework for composing SQL statements from R expressions using an S3 class hierarchy and non-standard evaluation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JakubBujnowicz/sqlqueries}
  \item Report bugs at \url{https://github.com/JakubBujnowicz/sqlqueries/issues}
}

}
\author{
\strong{Maintainer}: Jakub Bujnowicz \email{bujnowiczgithub@gmail.com}

Authors:
\itemize{
  \item Jakub Bujnowicz \email{bujnowiczgithub@gmail.com}
}

}
\keyword{internal}
