% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns_picker.R
\name{.columns_picker}
\alias{.columns_picker}
\title{Internal column picker helper}
\usage{
.columns_picker(..., .defuse = TRUE)
}
\arguments{
\item{...}{character vectors, names of columns. May also include other \code{sql} objects (e.g.\code{\link[=sql_vars]{sql_vars()}}, \code{\link[=sql_case]{sql_case()}}). Non-character elements are dropped with a warning, after dropping the vector cannot be empty. Vectors may be named to provide column aliases.}

\item{.defuse}{a logical value. Decides whether input arguments in \code{...}
should be defused. For more information about defusing, see \link{sql_defusing}.}
}
\value{
A character vector of column names.
}
\description{
DESCRIPTION TO BE WRITTEN
}
\keyword{internal}
