% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_logical.R
\name{.parse.sql_logical}
\alias{.parse.sql_logical}
\title{Internal parser for \code{sql_logical} objects}
\usage{
\method{.parse}{sql_logical}(x, fields, break_lines = TRUE, ...)
}
\arguments{
\item{x}{a \code{sql} object or a single string.}

\item{fields}{a list passed to \code{.parse} class methods. Internally set
within \code{.sql_parse()} wrapper to \code{attributes(x)}. The attributes
are set within object constructors.}

\item{break_lines}{a logical value. If \code{TRUE}, conditions are separated by
new lines instead of spaces.}

\item{...}{another arguments passed to \code{.parse()} and corresponding methods.}
}
\description{
Internal parser for \code{sql_logical} objects
}
\keyword{internal}
