% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_methods.R
\name{print.sql}
\alias{print.sql}
\title{Print SQL objects}
\usage{
\method{print}{sql}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Prints a SQL object to the console with a decorative header showing
the package name and a dashed line separator above and below the SQL
text. This method is automatically invoked when a \code{sql} object is
printed at the command line.
}
\examples{
print(sql_select("*"))

}
