% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_create_table.R
\name{sql_create_table}
\alias{sql_create_table}
\title{Construct a SQL CREATE TABLE statement}
\usage{
sql_create_table(name, variables = NULL, query = NULL)
}
\arguments{
\item{name}{a single string with the name of the table to be created.}

\item{variables}{a named character vector. Vector names represent names of
the columns, whereas vector values represent column types (e.g. integer).
Names must be unique.}

\item{query}{an \code{\link[=sql_query]{sql_query()}} object, query used as a basis
for the new table.}
}
\value{
A character string representing the SQL FROM statement, with S3
class 'sql_from'.
}
\description{
DESCRIPTION TO BE WRITTEN
}
