% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_delete.R
\name{sql_delete}
\alias{sql_delete}
\title{Construct a SQL DELETE statement}
\usage{
sql_delete(from, ..., where = NULL, .defuse = TRUE)
}
\arguments{
\item{from}{a single, non-empty string with a table name to delete from.}

\item{...}{a vector of conditions, passed to \code{\link[=sql_condition]{sql_condition()}}.}

\item{where}{a single \code{\link[=sql_where]{sql_where()}} object. Joined to conditions
from \code{...} with an AND statement. Can also be used on its own.}

\item{.defuse}{a logical value. Decides whether input arguments in \code{...}
should be defused. For more information about defusing, see \link{sql_defusing}.}
}
\value{
A character string representing the SQL DELETE statement, with S3
class 'sql_delete'.
}
\description{
Builds a DELETE statement to remove rows from a table. Conditions can
be passed directly as arguments or via a pre-built \code{\link[=sql_where]{sql_where()}} object.
If both are provided, they are combined with AND.
}
\examples{
sql_delete(from = "test_table", "a = 1")
wh <- sql_where("a > 3")
sql_delete(from = "test_table", where = wh)

}
\seealso{
SQL data manipulation functions:
\code{\link[=sql_create_table]{sql_create_table()}},
\code{\link[=sql_insert]{sql_insert()}},
\code{\link[=sql_update]{sql_update()}}
}
\concept{dml}
