% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_from.R
\name{sql_from}
\alias{sql_from}
\title{Construct a SQL FROM statement}
\usage{
sql_from(table, alias = NULL)
}
\arguments{
\item{table}{a single string with the name of the table. In particular, a
\code{sql_query} object may be passed as well, it is then automatically
wrapped in parentheses.}

\item{alias}{a single string or \code{NULL}. If provided, serves as an alias for
the table. Alias can also be provided by passing a named vector to
\code{table} (the name cannot be empty), however \code{alias} takes precedence.}
}
\value{
A character string representing the SQL FROM statement, with S3
class 'sql_from'.
}
\description{
Declares the source table for a query. Accepts a plain table name, a
named vector to define an alias, or a subquery (any \code{sql_query} object)
which is automatically wrapped in parentheses for use as a derived table.
}
\examples{
sql_from("dt")
sql_from("dt", alias = "t")
sql_from(c(t = "dt"))

}
\seealso{
SQL query statements:
\code{\link[=sql_group_by]{sql_group_by()}},
\code{\link[=sql_having]{sql_having()}},
\code{\link[=sql_join]{sql_join()}},
\code{\link[=sql_order_by]{sql_order_by()}},
\code{\link[=sql_select]{sql_select()}},
\code{\link[=sql_where]{sql_where()}}
}
\concept{building_blocks}
