% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_glue.R
\name{sql_glue}
\alias{sql_glue}
\title{Glue data into SQL objects}
\usage{
sql_glue(sql, .x, ...)
}
\arguments{
\item{sql}{a single string or a \code{sql} object.}

\item{.x}{[\code{listish}]\cr An environment, list, or data frame used to lookup values.}

\item{...}{[\code{expressions}]\cr Unnamed arguments are taken to be expression
string(s) to format. Multiple inputs are concatenated together before formatting.
Named arguments are taken to be temporary variables available for substitution.

For \code{glue_data()}, elements in \code{...} override the values in \code{.x}.}
}
\value{
A single string with \code{"{some_variable}"} text elements replaced.
If \code{sql} was an object of \code{sql} class, the class and associated attributes
are preserved.
}
\description{
A wrapper for \code{\link[glue:glue]{glue::glue_data()}}, which preserves attributes for \code{sql}
objects.
}
\details{
Note that underlying \code{\link[glue:glue]{glue::glue_data()}} is called with \code{.trim = FALSE}.
}
