% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_in.R
\name{sql_in}
\alias{sql_in}
\alias{\%IN\%}
\title{Construct a SQL IN statement}
\usage{
sql_in(x, vector)

x \%IN\% vector
}
\arguments{
\item{x}{a single string, usually a name of a variable}

\item{vector}{a single string or an atomic vector. Represents the set of
allowed values for the IN statement. If a vector of at least two elements
is provided, then \code{vector} is passed to \code{\link[=sql_tuple]{sql_tuple()}}.}
}
\value{
A character string representing the SQL IN statement, with S3
class 'sql_in'.
}
\description{
DESCRIPTION TO BE WRITTEN
}
\details{
\code{x \%IN\% y} is a wrapper for \code{sql_in(x, vector = y)}.
}
