% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_insert.R
\name{sql_insert}
\alias{sql_insert}
\title{Create a SQL INSERT statement}
\usage{
sql_insert(into, values = NULL, query = NULL, columns = NULL)
}
\arguments{
\item{into}{a single non-empty string, the name of a table to insert into.}

\item{values}{a data frame with at least one row, used as VALUES for
inserting. Values (columns) from the data frame are coerced into character
vectors (through \code{\link[=sql_tuple]{sql_tuple()}}).}

\item{query}{a \code{sql_query} object, may be used to insert through a select
query.}

\item{columns}{a character vector, can be used to insert into only a selected
subset of columns of \code{into} table.}
}
\value{
A character string representing the SQL INSERT statement, with S3
class 'sql_insert'.
}
\description{
Create a SQL INSERT statement
}
