% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_logical.R
\name{sql_logical}
\alias{sql_logical}
\alias{sql_condition}
\alias{sql_and}
\alias{sql_or}
\alias{\%AND\%}
\alias{\%OR\%}
\title{Internal function for creating logical conditions}
\usage{
sql_logical(x, y, operator)

sql_condition(..., .defuse = TRUE)

sql_and(x, y)

sql_or(x, y)

x \%AND\% y

x \%OR\% y
}
\arguments{
\item{x, y}{non-empty strings, input arguments for the logical operator.}

\item{operator}{a non-empty string, name of the logical operator to be
used (e.g. \code{"and"} or \code{"or"}).}

\item{...}{Arguments to be processed, representing conditions.}

\item{.defuse}{a logical value. Decides whether input arguments in \code{...}
should be defused. For more information about defusing, see \link{sql_defusing}.}
}
\value{
A character string representing the logical conditions, with S3
class 'sql_logical'.
}
\description{
This works as a template for creating logical conditions in a object-oriented
matterd, e.g. AND or OR.
}
