% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_and.R, R/sql_in.R, R/sql_logical.R,
%   R/sql_or.R
\name{sql_and}
\alias{sql_and}
\alias{\%AND\%}
\alias{sql_in}
\alias{\%IN\%}
\alias{sql_logical}
\alias{.new_logical}
\alias{sql_or}
\alias{\%OR\%}
\title{Title}
\usage{
sql_and(x, y)

x \%AND\% y

sql_in(x, vector)

x \%IN\% vector

.new_logical(x, y, operator)

sql_or(x, y)

x \%OR\% y
}
\arguments{
\item{operator}{}
}
\description{
Title
}
