% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_and.R, R/sql_logical.R, R/sql_or.R
\name{sql_and}
\alias{sql_and}
\alias{\%AND\%}
\alias{sql_logical}
\alias{sql_or}
\alias{\%OR\%}
\title{Title}
\usage{
sql_and(x, y, add_parenth = TRUE)

x \%AND\% y

sql_or(x, y, add_parenth = TRUE)

x \%OR\% y
}
\arguments{
\item{y}{}
}
\description{
Title
}
