% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_parenth.R
\name{sql_parenth}
\alias{sql_parenth}
\title{Wrap \code{sql} objects in parentheses}
\usage{
sql_parenth(x)
}
\arguments{
\item{x}{a single string, in particular any \code{sql} object.}
}
\value{
A character string wrapped in parentheses, with S3
class 'sql_parenth'.
}
\description{
DESCRIPTION TO BE WRITTEN
}
