% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_select.R
\name{sql_select}
\alias{sql_select}
\title{Construct a SQL SELECT statement}
\usage{
sql_select(
  ...,
  .distinct = FALSE,
  .top = NULL,
  .top_percent = FALSE,
  .top_with_ties = FALSE,
  .defuse = TRUE
)
}
\arguments{
\item{...}{character vectors, columns to select. May also include other \code{sql} objects (e.g.\code{\link[=sql_vars]{sql_vars()}}, \code{\link[=sql_case]{sql_case()}}). Non-character elements are dropped with a warning, after dropping the vector cannot be empty. Vectors may be named to provide column aliases.}

\item{.distinct}{a logical value. If \code{TRUE}, adds "DISTINCT" to the query.}

\item{.top}{a single number. If provided, adds a "TOP" clause.}

\item{.top_percent}{a logical value. If \code{TRUE}, adds "PERCENT" to the "TOP"
clause.}

\item{.top_with_ties}{a logical value. If \code{TRUE}, adds "WITH TIES" to the
"TOP" clause.}

\item{.defuse}{a logical value. Decides whether input arguments in \code{...}
should be defused. For more information about defusing, see \link{sql_defusing}.}
}
\value{
A character string representing the SQL SELECT statement, with S3
class 'sql_select'.
}
\description{
Specifies the columns to retrieve in a query. Accepts bare column names,
expressions, or aliased names. Supports DISTINCT for unique rows and TOP
for limiting results, including TOP PERCENT and WITH TIES modifiers.
}
\examples{
sql_select("a", "b")
sql_select("*", .distinct = TRUE)
sql_select("a", "b", .top = 5)

}
\seealso{
SQL query statements:
\code{\link[=sql_from]{sql_from()}},
\code{\link[=sql_group_by]{sql_group_by()}},
\code{\link[=sql_having]{sql_having()}},
\code{\link[=sql_join]{sql_join()}},
\code{\link[=sql_order_by]{sql_order_by()}},
\code{\link[=sql_where]{sql_where()}}
}
\concept{building_blocks}
