% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_tools.R
\name{sql_tree}
\alias{sql_tree}
\title{Convert \code{sql} objects to list trees}
\usage{
sql_tree(x)
}
\arguments{
\item{x}{a \code{sql} object.}
}
\value{
A list representing the tree.
}
\description{
This function can be used to explore internal structure of \code{sql} objects,
including data stored in object \code{fields} attribute. When used e.g. on \code{sql_query}
objects, the entire tree is returned as a nested list.
}
\examples{
sql_tree(sql_select("*"))
sql_tree(
    sql_select("*") +
        sql_from("dt"))

}
