% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_tuple.R
\name{sql_tuple}
\alias{sql_tuple}
\title{Create a SQL tuple}
\usage{
sql_tuple(...)
}
\arguments{
\item{...}{atomic vectors to create a tuple from. They all must be of
the same length.}
}
\value{
A single string with a SQL tuple.
}
\description{
DESCRIPTION TO BE WRITTEN
}
