% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_update.R
\name{sql_update}
\alias{sql_update}
\title{Construct a SQL UPDATE statement}
\usage{
sql_update(table, set)
}
\arguments{
\item{table}{a single string with the name of the table to update.}

\item{set}{a list, every element must be named (with unique names) and
must contain only a scalar value.}
}
\value{
A character string representing the SQL UPDATE statement, with S3
class 'sql_update'.
}
\description{
Builds an UPDATE statement to modify existing rows in a table. The
\code{set} parameter defines column-value pairs to assign. Character values
are quoted, missing values are rendered as NULL.
}
\examples{
sql_update(table = "test_table",
           set = list(b = "z"))
sql_update(table = "test_table",
           set = list(a = 1, b = "w"))

}
\seealso{
SQL data manipulation functions:
\code{\link[=sql_create_table]{sql_create_table()}},
\code{\link[=sql_delete]{sql_delete()}},
\code{\link[=sql_insert]{sql_insert()}}
}
\concept{dml}
