% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_tools.R
\name{sql_vars}
\alias{sql_vars}
\title{Prepare a variables list using Non-Standard Evaluation}
\usage{
sql_vars(...)
}
\arguments{
\item{...}{Expressions to convert into strings.}
}
\value{
A character vector of variables.
}
\description{
This function converts the provided expressions to character vectors, allowing
for easier inputs for e.g. SELECT statements. When an expression starts with
a minus operator, "DESC" is appended afterwards, which can be used for
easier typing in ORDER BY.
}
\examples{
sql_vars(Var1 = FirstVariable, Var2)
sql_vars(-FirstVariable, Var2)

}
