% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlqueries-package.R
\docType{package}
\name{sqlqueries-package}
\alias{sqlqueries}
\alias{sqlqueries-package}
\title{sqlqueries: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jakub Bujnowicz \email{bujnowiczgithub@gmail.com}

Authors:
\itemize{
  \item Jakub Bujnowicz \email{bujnowiczgithub@gmail.com}
}

}
\keyword{internal}
